% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot all accidents recorded a specified year in a state as black dots on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A state number as integer or character.}

\item{year}{A year as integer or character.}
}
\value{
A plot.
}
\description{
Plot all accidents recorded a specified year in a state as black dots on a map
}
\note{
Will throw error and not function if input state number
has no data points, if the year isn't related to a filename
or if several years / state.nums are set as inputs.
}
\examples{
\dontrun{
fars_map_state(13, 2013)
fars_map_state("4", "2015")
}
}
