% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load a csv file into a tbl data.frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character.}
}
\value{
A tbl data.frame.
}
\description{
Load a csv file into a tbl data.frame.
}
\note{
Will throw an error if file does not exist.
}
\examples{
\dontrun{
fars_read("mydata.csv")
fars_read("C:\\\\mydata.txt")
}
}
