% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load several accident_YEAR.csv.bz2 files into a list of tbl data.frame.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years (numeric or character).}
}
\value{
A list in which each element corresponds to a loaded file.
}
\description{
Load several accident_YEAR.csv.bz2 files into a list of tbl data.frame.
}
\note{
Will throw error and not function if any element inside years
isn't related to an existing filename.
}
\examples{
\dontrun{
fars_read_years(2013, 2014, 2015)
fars_read_years("2013")
}
}
