% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count the number of accidents for each month of every specified years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years (numeric or character).}
}
\value{
A tbl data.frame containing the summary (months in rows, years in columns)
}
\description{
Count the number of accidents for each month of every specified years
}
\note{
Will throw error and not function if any element inside years
isn't related to an existing filename.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years("2013")
}
}
