% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function
Accident Data set with year
This function gives the name of an accident dataset with "accident_year.csv.bz2" structure.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year of choice}
}
\value{
return to the chosen year with the dataset
}
\description{
Function
Accident Data set with year
This function gives the name of an accident dataset with "accident_year.csv.bz2" structure.
}
\examples{
makefile_name(2013)

}
