% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function
Map of accident.
This function indicates which location is prone to which accident}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{selected state}

\item{year}{selected year}
}
\value{
returns to map which shows the places where accident happned
}
\description{
Function
Map of accident.
This function indicates which location is prone to which accident
}
\examples{
fars_map_state(1,2014)
}
