% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate a "accident_YEAR.csv.bz2" string with a specified YEAR.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or a character.}
}
\value{
A character string.
}
\description{
Generate a "accident_YEAR.csv.bz2" string with a specified YEAR.
}
\note{
Will throw a warning and malfunction if year can't be
converted to an integer (eg. "HELLO" will produce
"accident_NA.csv.bz2").
}
\examples{
\dontrun{
make_filename(2014)
make_filename("2018")
}
}
